# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 工数見積もり API
  version: "1"
  description: |
    クライアントとサーバが共有する API。

    実装は `crates/api` にあり、**ローカルでもサーバでも同じコードが動く**。
    ローカルでは WASM の `api_call` から、サーバでは HTTP のハンドラから
    同じ `Service` を呼ぶため、権限の判定や不変条件が食い違うことがない。

    - 操作とルートの対応は `mhc_api::protocol::Request::route` が持っている。
      この文書はそれを人間向けに書き下したもので、両者が一致することは
      テストで確かめている。
    - 見積もりの計算 (`compute`) はこの API には含まれない。計算コアは
      クライアントに同梱された WASM がその場で回すので、サーバはデータの
      置き場所と権限だけを見る。
  license:
    name: MIT
servers:
  - url: /
    description: 同一オリジン
security:
  - bearerAuth: []
...
paths:
  /v1/me:
    get:
      tags: [me]
      summary: 自分の情報
      operationId: me
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/users:
    get:
      tags: [users]
      summary: アカウント一覧
      description: 共有先を選ぶために、管理者でなくても名前は見える。
      operationId: listUsers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [users]
      summary: アカウントを作る
      description: システム管理者のみ。
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, systemRole]
              properties:
                id: { type: string }
                name: { type: string, minLength: 1 }
                email: { type: [string, "null"], format: email }
                systemRole: { $ref: "#/components/schemas/SystemRole" }
      responses:
        "201":
          description: 作成した
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
  /v1/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/userId"
    patch:
      tags: [users]
      summary: アカウントを更新する
      description: |
        システム管理者、または本人。本人は自分の役割を変えられない。
        `clearEmail` を `true` にするとメールアドレスを空にする。
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1 }
                email: { type: string, format: email }
                clearEmail: { type: boolean, default: false }
                systemRole: { $ref: "#/components/schemas/SystemRole" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          $ref: "#/components/responses/Conflict"
        "422": { $ref: "#/components/responses/Invalid" }
    delete:
      tags: [users]
      summary: アカウントを削除する
      description: |
        システム管理者のみ。最後の管理者は消せない。
        どこかのプロジェクトの唯一の所有者である場合も、
        先に所有者を移すまで消せない (所有者のいないプロジェクトを作らないため)。
      operationId: deleteUser
      responses:
        "204": { description: 削除した }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/user-groups:
    get:
      tags: [user-groups]
      summary: アカウントのグループ一覧
      description: 共有先に選べるよう、管理者でなくても名前は見える。
      operationId: listUserGroups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [user-groups]
      summary: グループを作る
      description: システム管理者のみ。
      operationId: createUserGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name]
              properties:
                id: { type: string }
                name: { type: string, minLength: 1 }
      responses:
        "201":
          description: 作成した
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
  /v1/user-groups/{groupId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
    patch:
      tags: [user-groups]
      summary: グループを改名する
      description: システム管理者のみ。
      operationId: renameUserGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Invalid" }
    delete:
      tags: [user-groups]
      summary: グループを削除する
      description: |
        システム管理者のみ。このグループ経由でしか所有者が居ないプロジェクトが
        あるときは 409 で拒否する。付与されていた権限はすべて取り消される。
      operationId: deleteUserGroup
      responses:
        "204": { description: 削除した }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/user-groups/{groupId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
      - $ref: "#/components/parameters/userId"
    put:
      tags: [user-groups]
      summary: メンバーを入れる
      description: システム管理者のみ。すでに入っていれば何も起きない。
      operationId: addGroupMember
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [user-groups]
      summary: メンバーを外す
      description: システム管理者のみ。
      operationId: removeGroupMember
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/project-groups:
    get:
      tags: [project-groups]
      summary: プロジェクトのグループ一覧
      operationId: listProjectGroups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [project-groups]
      summary: グループを作る
      description: 作った本人が所有者になる。
      operationId: createProjectGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name]
              properties:
                id: { type: string }
                name: { type: string, minLength: 1 }
      responses:
        "201":
          description: 作成した
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
  /v1/project-groups/{groupId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
    patch:
      tags: [project-groups]
      summary: グループを改名する
      description: グループの所有者のみ。
      operationId: renameProjectGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Invalid" }
    delete:
      tags: [project-groups]
      summary: グループを削除する
      description: |
        グループの所有者のみ。配下のプロジェクトは消えず、所属だけが外れる。
        そのせいで所有者が居なくなるプロジェクトがあるときは 409 で拒否する。
      operationId: deleteProjectGroup
      responses:
        "204": { description: 削除した }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/project-groups/{groupId}/access/{principalKind}/{principalId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
      - $ref: "#/components/parameters/principalKind"
      - $ref: "#/components/parameters/principalId"
    put:
      tags: [project-groups]
      summary: グループに権限を与える・変更する
      description: |
        ここで与えた権限は、配下のすべてのプロジェクトに継がれる。
        グループの所有者が 1 人もいなくなる変更は 409 で拒否する。
      operationId: setGroupAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [project-groups]
      summary: グループの権限を取り消す
      operationId: removeGroupAccess
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        サーバ運用時のみ。ローカル (単一 HTML) では呼び出し元が常に
        そのブラウザの持ち主なので、認証は行われない。
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
    groupId:
      name: groupId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: 呼び出し元が特定できない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: 権限が足りない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: 対象が無い
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: 状態が合わない (最後の所有者を外そうとした、など)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Invalid:
      description: 入力が不正
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
...
    SystemRole:
      type: string
      enum: [admin, member]
      description: |
        `admin` は全プロジェクトとアカウントを管理できる。
        `member` は自分に共有されたプロジェクトだけを扱える。
...
    User:
      type: object
      required: [id, name, systemRole, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        systemRole: { $ref: "#/components/schemas/SystemRole" }
        createdAt: { type: string, format: date-time }
...